+++ xmake/modules/private/cache/build_cache.lua
@@ -232,6 +232,9 @@ function build(program, argv, opt)
         local objectfile_cached, objectfile_infofile = get(cachekey)
         if objectfile_cached then
             os.cp(objectfile_cached, cppinfo.objectfile)
+            -- we need update mtime for incremental compilation
+            -- @see https://github.com/xmake-io/xmake/issues/2620
+            os.touch(cppinfo.objectfile, {mtime = os.time()})
             -- we need get outdata/errdata to show warnings,
             -- @see https://github.com/xmake-io/xmake/issues/2452
             if objectfile_infofile and os.isfile(objectfile_infofile) then